PR changes incorporated
