+++ digiwf-spring-cloudstream-utils/src/main/java/io/muenchendigital/digiwf/spring/cloudstream/utils/api/streaming/MessageEmitter.java
@@ -9,7 +9,7 @@ import reactor.core.publisher.Sinks;
 import java.util.function.Supplier;
 
 @Configuration
-public class EventEmitter {
+public class MessageEmitter {
 
     @Bean
     public Sinks.Many<Message<Object>> sendMessageSink() {
